Make it possible to render DSL levels without translations
Right now, all DSL levels depend on I18n.t to render their content, and
will fall back to `nil` if they can't find the translation. We want to
be able to only sync data into the translation system if we want to
actually translate it, so we add the ability for DSL levels to fall back
to the database content if they can't find translation data.