Fix toggling between audio while playing
The way the InlineAudio element is currently implemented, if we try to
change the audio file to be played _while_ a file is currently being
played, then everything breaks.

Specifically, what happens is we detect that we are trying to update the
target file and unload the current one by unsetting the `src` attribute
of the Audio element and calling `load()` to make it unload itself.
Unfortunately, if this happens while the element is actively playing, it
will immediately try to play the `undefined` src value, catch the
inevitable error, and hide itself like it's supposed to (since we
usually interpret an error to mean no actual audio file was generated
for the given content).

To fix this, we simply make sure to `pause()` the current file before
unloading it, and also unload with `removeAttribute` rather than
assignment to undefined for clarity.