Add "My projects" menu item to user menu.
This works for both dashboard and pegasus. It goes to the same page as the "Projects" button,
as well as the "View all projects..." link at the bottom of the Create menu.

Also, this does not introduce the pegasus spreadsheet. If this seems to be the correct approach,
I can make the spreadsheet change first.