PL: code review feedback & test improvements
As well as code review feedback, this change unit tests the new priority deadline date field.

It was interesting to discover that the unit tests failed when the local machine had its time zone set to something other than PST, and it turns out to be a known issue discussed at https://github.com/travisjeffery/timecop/issues/182.  The solution is to use `Date.current` rather than `Date.today`.

A bunch of factories were also updated to use the yyyy-mm-dd date format that it turns out we're storing, as described in https://github.com/code-dot-org/code-dot-org/pull/27895