PL: fix for resending principal email for teacher application
This got slightly confusing: the teacher application model function got renamed from allow_resending_principal_email to allow_sending_principal_email since it's both used to tell the application dashboard that a principal email can be resent, but it's also used in the relevant API controller to check whether a principal email can be sent or resent from that same dashboard.  It turns out that the existing code in that dashboard relies on principal_approval_state, showing "send email" if we have nil for that, or allowing our new code to optionally show "resend email" if we have a value for that.