PL: script to queue reminder email to teacher awaiting principal approval
This script, not yet called as part of an automatic cron job, queues up a reminder email to a teacher awaiting principal approval for their teacher application.

One reason we queue here is so that we can examine the unsent emails before the daily cron job sends queued emails, giving us a chance to hand-inspect this behaviour on production as the feature is launched.

For an email to be queued, some criteria must be met: we must not have already created one of these reminder emails, the teacher application must be unreviewed or pending, the application must require but not yet have received principal approval, and it must have been at least 5 days since we last created a reminder email for the principal.  Except for the first one, this is the same set of requirements for showing a resend email button for the principal, as implemented in https://github.com/code-dot-org/code-dot-org/pull/28103