Support Blockly in new React-rendered markdown
Specifically, two changes are needed:

First, now that we are generating XML with React we end up with a bunch
of React comment nodes in our XML which break Blockly. Add a step to the
"render XML as Blockly" helper to clean out comment nodes.

Second, hast-util-sanitize by default clobbers 'name' attributes for
some reason. Blockly XML uses that attribute extensively, so let it
through.