Explicitly attempt to resolve i18n-specific and english-fallback documents as two separate steps
The current way we attempt to find localized templates leverages
extension names to make the localized version visible, and relies on
search ordering to make sure it gets selected first.

We'd like to implement some tweaks to the way we handle extension names
to allow for more dynamic template processing, but that becomes
difficult to do when the locale is also handled as an extension name.

With this change, we explicitly attempt to find the locale-specific
version of a file first by treating the locale as part of the
_filename_, not the extension name. This will both unblock us to make
tweaks to extension names and make the locale-vs-fallback functionality
more explicit.