Fix Inline Audio in non-english contained levels
Specifically, make the InlineAudio component only attempt to derive the
source file for a message string if it's actually given a message
string.

Right now, attempting to view a contained level with audis such as
https://studio.code.org/s/express-2018/stage/14/puzzle/13 breaks in
non-english locales because InlineAudio is attempting to perform an
operation on a nonexistent string. This is expected to happen in
situations where we don't have a translated audio string.