Sanitize XML
Sanitize XML in Markdown to allow only those tags and attributes we need
for Blockly.

Note that this list was generated based on an analysis of the Blockly
XML we _currently_ use, not on all the possible XML we _could_ use. It
could be that this list is inadequately permissive.