Fix: Workshops without subjects have exit surveys
In particular, workshops with a `NULL` subject will still be included in
the `with_surveys` enrollment scope, and therefore will remind about
pending exit surveys on the professional learning landing page.

Fixes a regression introduced in
https://github.com/code-dot-org/code-dot-org/pull/29511 when I added a
condition to the `with_surveys` scope, not realizing it would implicitly
exclude null values as well.

Did some work to evaluate impact, and it looks like we've had no
workshops with `NULL` subject end in the week or so since I introduced
this regression, so overall impact is near-zero.  Still, I'd like to get
it fixed before I move on to disabling exit surveys for Facilitator
workshops (which also have a `NULL` subject).