PL: Regional partner search shows less for closed workshops
When a regional partner is closed for workshop applications, show less.  In particular, the "Scholarship, discounts, and cost information:" heading and its grey box of content is not shown, and the "Additional program information:" heading and its content is also not shown.

Also, the "Or, apply directly now!" link that was permanently at the top of the page is permanently removed.