PL: make mailer test pass
It turns out that when we tell Mocha that we expect teacher_follow_up to be called, it doesn't call the real copy of that function any more.  This was causing an error in check_should_send which was still getting called, because there was no @workshop set beforehand.  By telling Mocha that we expect check_should_send to be called, it no longer calls the real copy of that function either, and the tests pass.