Fix pd-workshop-survey contact_id reference
This survey partial still contained a reference to
regional_partners.contact_id, which was removed in
https://github.com/code-dot-org/code-dot-org/pull/30445. Instead, I'm
using the workshop's regional_partner_id to do a regional partner
lookup.

This is intentionally a shallow fix, no cleanup or refactoring; I'm just
trying to quickly correct what looks like a regression on production.
We're hoping to deprecate this view soon anyway.