Donors: Teacher banner improvements
The banner can now be dismissed on pegasus, though it only works if the user is signed in.  The banner will also now be hidden on pegasus if the user is signed in and has previously dismissed it.

Note that the dashboard APIs responsible for this get and set will both return 403 errors if there is not a signed in user, but the user should not see anything amiss unless they're looking carefully at their browser developer tools.

Signed out users will now also see the banner on pegasus.