PL: Fix legacy regional partner contact form for non-US IPs
The legacy regional partner contact form at https://studio.code.org/pd/regional_partner_contact/new had a bug when the visitor's IP address was non-US.  The form tells its school-info partial to assume the visitor is in the US.  The partial hides the real country dropdown while the form provides its own which simply offers US vs. non-US.  The partial also attempts to set the real country dropdown to the "United States".  Unfortunately it should have been attempting to set it to "US" which is the actual value of that option.

Visitors geocoded to the US (or on localhost) didn't encounter this problem because separate initialization code inside the SchoolInfoManager set the real country dropdown to "US", but non-US visitors never got this rescue.

One result of this problem was that users choosing Public or Private schools weren't seeing the ZIP or state fields appear, because the code to make them visible also checks that country is "US".  I suspect this had a flow-on effect in which we weren't matching to regional partners for the automated emails because we didn't have the ZIP or state in the submitted contact forms.