Remove gsheet-based summing for state promote
Long ago (in https://github.com/code-dot-org/code-dot-org/pull/3591) we apparently took advantage of an extra row in the cdo-state-promote gsheet that was labeled with "Sum_states" rather than a state itself.  This caused an issue in the iterator at https://github.com/code-dot-org/code-dot-org/blob/staging/pegasus/rake/generate_pdfs.rake#L21 when we attempted to generate a PDF for a state of "Sum_states".

With this change, we just do the summing in the SQL query, rather than relying upon the sum generated by the gsheet.  This will allow us to simultaneously remove the row from the gsheet.