Destroy resolved QueuedAccountPurges
Due to an increase in intermittent Pardot errors, we recently [made some
QueuedAccountPurges
autoretryable](https://github.com/code-dot-org/code-dot-org/pull/30909).
(See also [this change to reporting from this
task](https://github.com/code-dot-org/code-dot-org/pull/30959).) To our
surprise, the queue depth still continued to increase.

It turns out the auto-retry was working, but the `QueuedAccountPurge`
records associated with the retried accounts were not being cleared.

This change adds a step right at the end of the nightly task to clean up
any resolved `QueuedAccountPurge` records.  One benefit of this design
is that it will retroactively clean up our queue on production next time
it runs, so no manual correction to the production database is required.