PL: Convert teacher application MarkdownSpan -> SafeMarkdown
Back in https://github.com/code-dot-org/code-dot-org/pull/27841 we attempted converting the markdown-based labels in the teacher application from MarkdownSpan to UnsafeRenderedMarkdown (which is happily now known as SafeMarkdown as of https://github.com/code-dot-org/code-dot-org/pull/29457).  However, the red asterisks for required fields were apparently not showing in the right place, as evidenced by the work in https://github.com/code-dot-org/code-dot-org/pull/28152 to move them to the fronts of the questions.

This change converts the same labels from MarkdownSpan to SafeMarkdown, but this time leaves the red asterisks at the ends of the labels.  It does this by wrapping the label in a div named for whether the field is required or not, and then we use a little CSS trick to put the red asterisk at the end of the last <p> tag generated by the markdown for that label.