hoc2019: Support for Minecraft: Education Edition certificate
This supports showing a Minecraft: Education Edition certificate when that tutorial is complete.  It's identified by tutorial code "mee", and this scenario should be tested once it has been entered in the cdo-tutorials gsheet.  Because we don't host it ourselves, we special-case this certificate's handling.

The certificate image is actually used twice: once by apps code when showing the certificate initially at https://studio.code.org/congrats; and then by pegasus code when it is customised with a user-entered name during an ajax call to https://code.org/api/hour/certificate/[filename].