Default notifications to Liz when partner is missing
Application notification emails that would normally go to our regional
partners should instead be sent to Liz when no partner was matched for a
given application.  At this time there's only one such notification
(that a principal approval form was completed) and we were simply
skipping it when no partner was found.