Remove Pd::CsfCertificateController
Deletes Pd::CsfCertificateController and removes the
`/generate_csf_certificate/:code` route that was [deprecated in June of
this year][0].

We've been monitoring traffic to this route [via Honeybadger][1] and
have only seen five hits in the last four months.  That's a volume we
are happy to handle through our support desk if needed.

[0]: https://github.com/code-dot-org/code-dot-org/pull/28762
[1]: https://app.honeybadger.io/projects/3240/faults/50875900