Fixing /yourschool Mapbox size and button size
The Mapbox map on code.org/yourschool had a fixed size which made it
troublesome to use on small screens. The size of the map now scales
appropriately to the size of the screen.

Some CSS style names were also refactored to match best practices.

The margin around the Mapbox buttons were also fixed.