Handle submissionless peer reviews
Test and fix a relatively rare case with Peer Reviews where a UserLevel
exists but no PeerReviews do yet.  This is worth covering and supporting
if only because we've had so many data issues with this system in the
past.

I've tried to construct a plausible scenario for testing, although I'm
not sure this situation is easy to get into with our UI as it currently
exists.

This should address [this Honeybadger error][hb].

[hb]: https://app.honeybadger.io/projects/3240/faults/57023032