Fix continue button on external markdown levels
Fixes the continue button on external markdown levels.  This regressed
in https://github.com/code-dot-org/code-dot-org/pull/31726 when we were
trying to fix the continue button for external video levels, but we
swapped in a jQuery expression that removed the dependency on a
particular static parent, but also removed the lazy-binding feature that
markdown levels were depending on (the submit button might be added
after this code runs in external markdown levels).

As a solution, I'm putting back the lazy-binding code with a more
generic static parent that's present in both level types.  I've also
added a regression test using an existing allthethings level the
demonstrated this problem.

Raised [in
Slack](https://codedotorg.slack.com/archives/CA3KCSGTD/p1573765609181600)
by @joshlory