Fix syncing block_categories
In https://github.com/code-dot-org/code-dot-org/pull/29117, I refactored
the sync in process to organize content by level rather than by content
type. As part of that, I also updated `block_categories` to be
serialized into JSON rather than YAML.

Although this change meant we were using a better data serialization
format, it unfortunately is not the format recognized by either rails or
the part of the sync out that distributes rails content:
https://github.com/code-dot-org/code-dot-org/blob/7460a3f7788cf1d19a408b280a076c76472347de/bin/i18n/sync-out.rb#L208-L218

One option for fixing this would be to update the sync out to convert
this content into a format recognized by rails, but in the long term we
would like to update the rails i18n engine itself to be smarter about
the format it expects content to be in. Work to update the sync out
would be wasted in that context, so for this item I prefer to simply
update the sync in to create `block_categories` in the desired format
right off the bat.

Follow-up:

After running a sync with this change, the existing
`block_categories.json` file should be removed from both crowdin and
this repo.