Oceans: fix rotation image
In the new Oceans script, the device rotation image shown on portrait layout was not showing properly.  This is a fix for that issue targeted at the Oceans script, but it could be generalised later, since this apparently happens for other tutorials too.

On an actual iPhone, this view is scrollable.  In Chrome's device mode it's not, but the issue is otherwise the same.  The issue here is that the viewport's dimensions and scaling have made the viewport bigger than the physical screen.

The fix is fairly simple in this case.  We know that the Oceans tutorial can scale down to 480px wide, and so we set the viewport to be 600px wide.  We also set the scaling on the viewport so that the 600px is what's displayed on the screen.  We also adjust the scaling on the rotation image to fill that space.