Fix fallback full-width video player layout
https://github.com/code-dot-org/code-dot-org/pull/31797 added support for full-width videos, notably standalone video levels which feature a video player that expands responsively to fill the screen.  However, the layout was broken for the fallback video player.

This fix gets the fallback player filling the same space as the YouTube video would.  It's fairly straightforward: in the case of a "full-width" video, our code sets its width and height to be "100%" rather than a specific pixel width.  Then, knowing that its host is a "position: relative" div of the desired location and dimensions, an intermediate div marks itself as "position: absolute" and extending to the four bounding edges of that parent.  The fallback player video is attached to that, as before.

Being so close to Hour of Code 2019, manual testing has been done to ensure no regressions.  In particular, the following levels have been loaded up and all appear to work as before:

/s/oceans/stage/1/puzzle/1 (full-width standalone with YouTube)
/s/allthethings/stage/34/puzzle/1 (regular standalone with YouTube)
/s/allthethings/stage/34/puzzle/1?force_youtube_fallback=1 (regular standalone with fallback player)
/s/allthethings/stage/2/puzzle/1 (maze level with pop-up Youtube)
/s/allthethings/stage/2/puzzle/1?force_youtube_fallback (maze level with pop-up fallback player)