Adding Blockly blocks to Content levels.
In Content levels, such as External levels, level creators were taking
screenshots of Blockly blocks and adding them to the instructions in
order to give a preview to the student of what the blocks look like.
This has become a problem because screenshots are not translatable and
we are expanding our content to be used internationally.
* Added common, maze, and turtle CDO blocks to Content levels.
* <xml> blocks in Markdown content render as Blockly blocks.