Restore created_at on workshop dashboard
Restores behavior introduced in
https://github.com/code-dot-org/code-dot-org/pull/32453 and regressed
during a complicated rebase in
https://github.com/code-dot-org/code-dot-org/pull/31990.  Most of the
change was preserved, but the JSX that actually renders the new
component was accidentally dropped.

I mentioned in the previous PR that I wasn't adding a test for the UI
because I was adding so many tests in the other PR (that regressed
this!) and it would be covered.  Turns out I was wrong!  So now I'm
adding a unit test for the Workshop component that verifies that we show
the created_at date as expected.