Video players: allow force youtube URL parameter
The dashboard and pegasus video players already support a `force_youtube_fallback` URL parameter to force the fallback player to run.

Now, they also support a `force_youtube_player` URL parameter to force YouTube to run.

This might be useful for diagnosing video player issues in the future, for example if the fallback detection kicks in too aggressively because the user is on a particularly slow connection.