PL: CSF workshop optional fee text - migration
Add serialized attributes to Pd::Workshop, along with a first use of them: an optional fee text for CSF workshops.

The use of serialized attributes will save us from having to create a new migration (and new database column) every time we want to add some metadata like this to a workshop.