Fixing teacher name size on workshop certs
A recent change was made to make the font sizes on certificates to be
dynamically sized. Most of the certificates have the student or
teacher's name in about the same spot at the same size. However,
workshop certificates are a much higher resolution than the other
certificates. The teacher's name appeared small on the certificate
because the bounding box for the text is defined in pixels and it is too
small for the higher resolution workshop certificate. This change
increases the size of the bounding boxes for text on workshop
certificates. Bounding boxes were also checked for all text fields on
all certificates.

* Add width/height for all the text on the workshop certificates
* Made all certificates support slightly longer student/teacher names.