PL: generate legacy survey summaries
This script can generate a variety of summaries and write them to the Pd::LegacySurveySummary introduced in https://github.com/code-dot-org/code-dot-org/pull/33451.

As we transition away from pegasus- and jotform-based surveys, we also want to deprecate the code that generates and displays their summaries, but we've been asked to still display legacy averages per faciitator.

The script is called as follows:

  Usage: generate_legacy_survey_summaries.rb
             csf_intro_post_workshop |
             csf_intro_post_workshop_all_workshops |
             csd_summer_workshops |
             csp_summer_workshops

It will delete existing entries for the specified category before writing new ones.  It should be used with caution once we've taken final snapshots.  But it should be safe to use repeatedly during development and testing.