fix 'finish' button to show on small screens for most free play levels
To resolve LP-1104. The finish button has been observed to not show up on free
play levels on small resolution screens. This commit contains a fix for most
free play levels on most devices, but there are still some issues that will
need to be resolved Bounce, minecraft, and app lab all fail to show the finish
button in the viewport under some circumstances, outlined in comments for end
to end tests. The scss fix was made by Brendan Reville