Commit content: notify DOTD when skip or fail
When attempt to commit content, usually as part of a cron job, now notify the Developer of the Day when it's skipped or failed.

For an example of how this should be helpful: sometimes the commit is skipped because the Slack channel's topic was set to "DTL: no" earlier, and it's easy for the DOTD to miss this.