Commit content: only notify DOTD on skip
Followup to https://github.com/code-dot-org/code-dot-org/pull/33972.  It turns out the failure case already notified the DOTD, and the string modified actually inadvertently added the DOTD's handle to the channel topic when we didn't want that.  This leaves the other change intact, which is to add the DOTD' tag to the chat message generated when content is skipped, which was the original goal of the PR anyway.