Mobile zoom: improve the "rotate phone" UI
This makes a substantial improvement to the "rotate phone" overlay, by using the newer Visual Viewport API, when available, to dynamically position and resize the content of the overlay to always be in view, even as the user adjusts zoom or position on their phone.

This also adjusts the viewport to better suit an Android issue we saw, and is described in a comment, and removes the special parameter for setting up the "rotate phone" overlay for the Oceans tutorial.