Add validation exception for old teacher accounts.
A few years ago, we added an email requirement for teacher accounts.
At the time, we did not backfill existing teacher accounts, so we still
have active teacher accounts which don't have emails recorded for them.
These users currently experience pain because the code currently refuses
to update change to their account because the account is currently
invalid. This change makes it so these older accounts will be considered
valid for the time being. There is follow up work to get an email
recorded for the teacher accounts.